Split up fetch.py into submodules.
