+++ pyproject.toml
@@ -30,7 +30,6 @@ disallow_untyped_calls = true
 disallow_untyped_decorators = true
 disallow_untyped_defs = true
 ignore_missing_imports = true
-implicit_reexport = false
 no_implicit_optional = true
 strict_equality = true
 warn_redundant_casts = true
