+++ src/fuzzfetch/__main__.py
@@ -4,6 +4,6 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from .fetch import Fetcher
+from . import Fetcher
 
 Fetcher.main()
