+++ src/fuzzfetch/core.py
@@ -5,27 +5,24 @@
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
 import configparser
-import itertools
 import logging
 import os
 import platform as std_platform
 import re
 import shutil
 import tempfile
-import time
-from argparse import SUPPRESS, ArgumentParser, Namespace
-from collections import namedtuple
 from datetime import datetime, timedelta
-from enum import Enum
 from pathlib import Path
 from sys import version_info
-from typing import Any, Dict, Iterable, Iterator, List, Optional, Sequence, Tuple, Union
+from typing import Any, Dict, Optional, Sequence, Tuple, Union
 
 from pytz import timezone
-from requests import Response, Session
-from requests.exceptions import RequestException
 
+from . import FetcherException
+from .args import FetcherArgs
+from .download import download_url, get_url, resolve_url
 from .extract import extract_dmg, extract_tar, extract_zip
+from .models import BuildFlags, BuildSearchOrder, BuildTask, HgRevision, Platform
 from .path import PathArg, onerror
 from .path import rmtree as junction_rmtree
 
@@ -37,21 +34,6 @@ else:
     from importlib.metadata import PackageNotFoundError, version
 
 
-__all__ = (
-    "__version__",
-    "BuildFlags",
-    "BuildSearchOrder",
-    "BuildTask",
-    "Fetcher",
-    "FetcherArgs",
-    "FetcherException",
-    "Platform",
-    "get_url",
-    "download_url",
-    "iec",
-    "si",
-)
-
 try:
     __version__ = version("fuzzfetch")
 except PackageNotFoundError:
@@ -59,692 +41,13 @@ except PackageNotFoundError:
     __version__ = None
 
 LOG = logging.getLogger("fuzzfetch")
-
-
 BUG_URL = "https://github.com/MozillaSecurity/fuzzfetch/issues/"
-HTTP_SESSION = Session()
-
-
-class FetcherException(Exception):
-    """Exception raised for any Fetcher errors"""
-
-
-def iec(number: Union[float, int]) -> str:
-    """Format a number using IEC multi-byte prefixes.
-
-    Arguments:
-        number: Number to format.
-
-    Returns:
-        Input number, formatted to the largest whole SI prefix.
-    """
-    prefixes = ["", "Ki", "Mi", "Gi", "Ti", "Pi", "Ei", "Zi", "Yi"]
-    while number > 1024:
-        number /= 1024.0
-        prefixes.pop(0)
-    return f"{number:0.2f}{prefixes[0]}"
-
-
-def si(number: Union[float, int]) -> str:  # pylint: disable=invalid-name
-    """Format a number using SI prefixes.
-
-    Arguments:
-        number: Number to format.
-
-    Returns:
-        Input number, formatted to the largest whole SI prefix.
-    """
-    prefixes = ["", "k", "M", "G", "T", "P", "E", "Z", "Y"]
-    while number > 1000:
-        number /= 1000.0
-        prefixes.pop(0)
-    return f"{number:0.2f}{prefixes[0]}"
-
-
-def get_url(url: str) -> Response:
-    """Retrieve requested URL"""
-    try:
-        data = HTTP_SESSION.get(url, stream=True)
-        data.raise_for_status()
-    except RequestException as exc:
-        raise FetcherException(exc) from None
-
-    return data
-
-
-def resolve_url(url: str) -> Response:
-    """Resolve requested URL"""
-    try:
-        data = HTTP_SESSION.head(url)
-        data.raise_for_status()
-    except RequestException as exc:
-        raise FetcherException(exc) from None
-
-    return data
-
-
-class HgRevision:
-    """Class representing a Mercurial revision."""
-
-    def __init__(self, revision: str, branch: str) -> None:
-        """Create a Mercurial revision object.
-
-        Arguments:
-            revision: revision hash (short or long)
-            branch: branch where revision is located
-        """
-        if branch is None or branch == "?":
-            raise FetcherException(f"Can't lookup revision date for branch: {branch}")
-
-        if branch == "autoland":
-            branch = f"integration/{branch}"
-        elif branch in {"release", "beta"} or branch.startswith("esr"):
-            branch = f"releases/mozilla-{branch}"
-        elif branch != "try":
-            branch = f"mozilla-{branch}"
-        self._data = get_url(
-            f"https://hg.mozilla.org/{branch}/json-rev/{revision}"
-        ).json()
-
-    @property
-    def pushdate(self) -> datetime:
-        """Get datetime object representing pushdate of the revision."""
-        push_date = datetime.fromtimestamp(self._data["pushdate"][0])
-        # For some reason this timestamp is always EST despite saying it has an UTC
-        # offset of 0.
-        return timezone("EST").localize(push_date)
-
-    @property
-    def hash(self) -> str:
-        """Get the long hash of the revision."""
-        assert isinstance(self._data["node"], str)
-        return self._data["node"]
-
-
-def download_url(url: str, outfile: PathArg) -> None:
-    """Download a URL to a local path.
-
-    Arguments:
-        url: URL to download.
-        outfile: Path to output file.
-    """
-    downloaded = 0
-    start_time = report_time = time.time()
-    resp = get_url(url)
-    total_size = int(resp.headers["Content-Length"])
-    LOG.info("> Downloading: %s (%sB total)", url, iec(total_size))
-    with open(outfile, "wb") as build_zip:
-        for chunk in resp.iter_content(1024 * 1024):
-            build_zip.write(chunk)
-            downloaded += len(chunk)
-            now = time.time()
-            if (now - report_time) > 30 and downloaded != total_size:
-                LOG.info(
-                    ".. still downloading (%0.1f%%, %sB/s)",
-                    100.0 * downloaded / total_size,
-                    si(float(downloaded) / (now - start_time)),
-                )
-                report_time = now
-    LOG.info(
-        ".. downloaded (%sB/s)", si(float(downloaded) / (time.time() - start_time))
-    )
 
 
 def _create_utc_datetime(datetime_string: str) -> datetime:
     """Convert build_string to time-zone aware datetime object"""
     dt_obj = datetime.strptime(datetime_string, "%Y%m%d%H%M%S")
     return timezone("UTC").localize(dt_obj)
-
-
-class BuildFlags(
-    namedtuple(
-        "BuildFlagsBase",
-        (
-            "asan",
-            "tsan",
-            "debug",
-            "fuzzing",
-            "coverage",
-            "valgrind",
-            "no_opt",
-            "fuzzilli",
-        ),
-    )
-):
-    """Class for storing TaskCluster build flags"""
-
-    def build_string(self) -> str:
-        """
-        Taskcluster denotes builds in one of two formats:
-        i.e. linux64-asan or linux64-asan-opt
-        The latter is generated. If it fails, the caller should try the former.
-        """
-        return (
-            ("-ccov" if self.coverage else "")
-            + ("-fuzzilli" if self.fuzzilli else "")
-            + ("-fuzzing" if self.fuzzing else "")
-            + ("-asan" if self.asan else "")
-            + ("-tsan" if self.tsan else "")
-            + ("-valgrind" if self.valgrind else "")
-            + ("-noopt" if self.no_opt else "")
-            + ("-debug" if self.debug else "")
-            + ("-opt" if not self.no_opt and not self.debug else "")
-        )
-
-
-class Platform:
-    """Class representing target OS and CPU, and how it maps to a Gecko mozconfig"""
-
-    SUPPORTED = {
-        "Darwin": {"x86_64": "macosx64"},
-        "Linux": {"x86_64": "linux64", "x86": "linux"},
-        "Windows": {"x86_64": "win64", "arm64": "win64-aarch64"},
-        "Android": {
-            "x86_64": "android-x86_64",
-            "x86": "android-x86",
-            "arm": "android-api-16",
-            "arm64": "android-aarch64",
-        },
-    }
-    CPU_ALIASES = {
-        "ARM64": "arm64",
-        "AMD64": "x86_64",
-        "aarch64": "arm64",
-        "i686": "x86",
-        "x64": "x86_64",
-    }
-
-    def __init__(
-        self, system: Optional[str] = None, machine: Optional[str] = None
-    ) -> None:
-        if system is None:
-            system = std_platform.system()
-        if machine is None:
-            machine = std_platform.machine()
-        if system not in self.SUPPORTED:
-            raise FetcherException(f"Unknown system: {system}")
-        fixed_machine = self.CPU_ALIASES.get(machine, machine)
-        if fixed_machine not in self.SUPPORTED[system]:
-            raise FetcherException(f"Unknown machine for {system}: {machine}")
-        self.system = system
-        self.machine = fixed_machine
-        self.gecko_platform = self.SUPPORTED[system][fixed_machine]
-
-    @classmethod
-    def from_platform_guess(cls, build_string: str) -> "Platform":
-        """Create a platform object from a namespace build string"""
-        match: List[str] = []
-        for system, platform in cls.SUPPORTED.items():
-            for machine, platform_guess in platform.items():
-                if platform_guess in build_string and (
-                    not match or len(match[2]) < len(platform_guess)
-                ):
-                    match = [system, machine, platform_guess]
-        if match:
-            return cls(match[0], match[1])
-        raise FetcherException(f"Could not extract platform from {build_string}")
-
-    def auto_name_prefix(self) -> str:
-        """
-        Generate platform prefix for cross-platform downloads.
-        """
-        # if the platform is not native, auto_name would clobber native downloads.
-        # make a prefix to avoid this
-        native_system = std_platform.system()
-        native_machine = self.CPU_ALIASES.get(
-            std_platform.machine(), std_platform.machine()
-        )
-        if native_system == self.system and native_machine == self.machine:
-            return ""
-        platform = {
-            "linux": "linux32",
-            "android-api-16": "android-arm",
-            "android-aarch64": "android-arm64",
-        }.get(self.gecko_platform, self.gecko_platform)
-        return f"{platform}-"
-
-
-class BuildTask:
-    """Class for storing TaskCluster build information"""
-
-    TASKCLUSTER_API = "https://firefox-ci-tc.services.mozilla.com/api/%s/v1"
-    RE_DATE = re.compile(r"^\d{4}-\d{2}-\d{2}$")
-    RE_REV = re.compile(r"^([0-9A-F]{12}|[0-9A-F]{40})$", re.IGNORECASE)
-
-    def __init__(
-        self,
-        build: Optional[str],
-        branch: Optional[str],
-        flags: Optional[BuildFlags],
-        platform: Optional[Platform] = None,
-        _blank: bool = False,
-    ) -> None:
-        """Retrieve the task JSON object
-
-        Requires first generating the task URL based on the specified build type and
-        platform
-        """
-        if _blank:
-            self.url: Optional[str] = None
-            self.queue_server: Optional[str] = None
-            self._data: Dict[str, Any] = {}
-            return
-        assert build is not None
-        assert branch is not None
-        assert flags is not None
-        for obj in self.iterall(build, branch, flags, platform=platform):
-            self.url = obj.url
-            self.queue_server = obj.queue_server
-            self._data = obj._data  # pylint: disable=protected-access
-            break
-        else:
-            raise FetcherException(
-                f"Unable to find usable archive for {self._debug_str(build)}"
-            )
-
-    @classmethod
-    def _debug_str(cls, build: str) -> str:
-        if cls.RE_DATE.match(build):
-            return f"pushdate {build}"
-        if cls.RE_REV.match(build):
-            return f"revision {build}"
-        return build
-
-    @classmethod
-    def iterall(
-        cls,
-        build: str,
-        branch: str,
-        flags: BuildFlags,
-        platform: Optional[Platform] = None,
-    ) -> Iterator["BuildTask"]:
-        """Generator for all possible BuildTasks with these parameters"""
-        # Prepare build type
-        if platform is None:
-            platform = Platform()
-        target_platform = platform.gecko_platform
-
-        is_namespace = False
-        if cls.RE_DATE.match(build):
-            flag_str = flags.build_string()
-            task_template_paths: Iterable[Tuple[str, str]] = tuple(
-                (template, path + flag_str)
-                for (template, path) in cls._pushdate_template_paths(
-                    build.replace("-", "."), branch, target_platform
-                )
-            )
-
-        elif cls.RE_REV.match(build):
-            # If a short hash was supplied, resolve it to a long one.
-            if len(build) == 12:
-                build = HgRevision(build, branch).hash
-            flag_str = flags.build_string()
-            task_paths = tuple(
-                path + flag_str
-                for path in cls._revision_paths(build.lower(), branch, target_platform)
-            )
-            task_template_paths = itertools.product((cls.TASKCLUSTER_API,), task_paths)
-
-        elif build == "latest":
-            if branch not in {"autoland", "try"}:
-                branch = f"mozilla-{branch}"
-
-            if not any(flags):
-                # Opt builds are now indexed under 'shippable'
-                namespace = f"gecko.v2.{branch}.shippable.latest"
-            else:
-                namespace = f"gecko.v2.{branch}.latest"
-
-            prod = "mobile" if "android" in target_platform else "firefox"
-            task_paths = (
-                f"/task/{namespace}.{prod}.{target_platform}{flags.build_string()}",
-                f"/task/{namespace}.{prod}.sm-{target_platform}{flags.build_string()}",
-            )
-            task_template_paths = itertools.product((cls.TASKCLUSTER_API,), task_paths)
-
-        else:
-            # try to use build argument directly as a namespace
-            task_path = f"/task/{build}"
-            is_namespace = True
-            task_template_paths = ((cls.TASKCLUSTER_API, task_path),)
-
-        for (template_path, try_wo_opt) in itertools.product(
-            task_template_paths, (False, True)
-        ):
-
-            template, path = template_path
-
-            if try_wo_opt:
-                if "-opt" not in path or is_namespace:
-                    continue
-                path = path.replace("-opt", "")
-
-            try:
-                url = (template % ("index",)) + path
-                data = HTTP_SESSION.get(url)
-                data.raise_for_status()
-            except RequestException:
-                continue
-
-            obj = cls(None, None, None, _blank=True)
-            obj.url = url
-            obj.queue_server = template % ("queue",)
-            obj._data = data.json()  # pylint: disable=protected-access
-
-            LOG.debug("Found archive for %s", cls._debug_str(build))
-            yield obj
-
-    def __getattr__(self, name: str) -> Any:
-        if name in self._data:
-            return self._data[name]
-        raise AttributeError(
-            f"'{type(self).__name__}' object has no attribute '{name}'"
-        )
-
-    @classmethod
-    def _pushdate_template_paths(
-        cls, pushdate: str, branch: str, target_platform: str
-    ) -> Iterator[Tuple[str, str]]:
-        """Multiple entries exist per push date. Iterate over all until a working entry
-        is found
-        """
-        if branch not in {"autoland", "try"}:
-            branch = f"mozilla-{branch}"
-
-        paths = (
-            f"/namespaces/gecko.v2.{branch}.shippable.{pushdate}",
-            f"/namespaces/gecko.v2.{branch}.pushdate.{pushdate}",
-        )
-
-        for path in paths:
-            index_base = cls.TASKCLUSTER_API % ("index",)
-            url = index_base + path
-            try:
-                base = HTTP_SESSION.post(url, json={})
-                base.raise_for_status()
-            except RequestException:
-                continue
-
-            product = "mobile" if "android" in target_platform else "firefox"
-            json = base.json()
-            for namespace in sorted(json["namespaces"], key=lambda x: str(x["name"])):
-                task_paths = (
-                    f"/task/{namespace['namespace']}.{product}.{target_platform}",
-                    f"/task/{namespace['namespace']}.{product}.sm-{target_platform}",
-                )
-                for pair in itertools.product((cls.TASKCLUSTER_API,), task_paths):
-                    yield pair
-
-    @classmethod
-    def _revision_paths(
-        cls, rev: str, branch: str, target_platform: str
-    ) -> Iterator[str]:
-        """Retrieve the API path for revision based builds"""
-        if branch not in {"autoland", "try"}:
-            branch = f"mozilla-{branch}"
-
-        namespaces = (
-            f"gecko.v2.{branch}.revision.{rev}",
-            f"gecko.v2.{branch}.shippable.revision.{rev}",
-        )
-
-        for namespace in namespaces:
-            product = "mobile" if "android" in target_platform else "firefox"
-            yield f"/task/{namespace}.{product}.{target_platform}"
-            yield f"/task/{namespace}.{product}.sm-{target_platform}"
-
-
-class FetcherArgs:
-    """Class for parsing and recording Fetcher arguments"""
-
-    DEFAULT_TARGETS = ["firefox"]
-
-    def __init__(self) -> None:
-        """Instantiate a new FetcherArgs instance"""
-        super().__init__()  # call super for multiple-inheritance support
-        if not hasattr(self, "parser"):
-            self.parser = ArgumentParser(conflict_handler="resolve", prog="fuzzfetch")
-
-        self.parser.set_defaults(
-            target="firefox", build="latest", tests=None
-        )  # branch default is set after parsing
-
-        target_group = self.parser.add_argument_group("Target")
-        target_group.add_argument(
-            "--target",
-            nargs="*",
-            default=FetcherArgs.DEFAULT_TARGETS,
-            help="Specify the build artifacts to download. "
-            "Valid options: firefox js common gtest "
-            f"(default: {' '.join(FetcherArgs.DEFAULT_TARGETS)})",
-        )
-        target_group.add_argument(
-            "--os",
-            choices=sorted(Platform.SUPPORTED),
-            help=f"Specify the target system. (default: {std_platform.system()})",
-        )
-        cpu_choices = sorted(
-            set(
-                itertools.chain(
-                    itertools.chain.from_iterable(Platform.SUPPORTED.values()),
-                    Platform.CPU_ALIASES,
-                )
-            )
-        )
-        target_group.add_argument(
-            "--cpu",
-            choices=cpu_choices,
-            help=f"Specify the target CPU. (default: {std_platform.machine()})",
-        )
-
-        type_group = self.parser.add_argument_group("Build")
-        type_group.add_argument(
-            "--build",
-            metavar="DATE|REV|NS",
-            help="Specify the build to download, (default: %(default)s)"
-            " Accepts values in format YYYY-MM-DD (2017-01-01)"
-            " revision (57b37213d81150642f5139764e7044b07b9dccc3)"
-            " or TaskCluster namespace (gecko.v2....)",
-        )
-
-        branch_group = self.parser.add_argument_group("Branch")
-        branch_args = branch_group.add_mutually_exclusive_group()
-        branch_args.add_argument(
-            "--central",
-            action="store_const",
-            const="central",
-            dest="branch",
-            help="Download from mozilla-central (default)",
-        )
-        branch_args.add_argument(
-            "--release",
-            action="store_const",
-            const="release",
-            dest="branch",
-            help="Download from mozilla-release",
-        )
-        branch_args.add_argument(
-            "--beta",
-            action="store_const",
-            const="beta",
-            dest="branch",
-            help="Download from mozilla-beta",
-        )
-        branch_args.add_argument(
-            "--esr-stable",
-            action="store_const",
-            const="esr-stable",
-            dest="branch",
-            help="Download from esr-stable",
-        )
-        branch_args.add_argument(
-            "--esr-next",
-            action="store_const",
-            const="esr-next",
-            dest="branch",
-            help="Download from esr-next",
-        )
-        branch_args.add_argument(
-            "--try",
-            action="store_const",
-            const="try",
-            dest="branch",
-            help="Download from try",
-        )
-        branch_args.add_argument(
-            "--autoland",
-            action="store_const",
-            const="autoland",
-            dest="branch",
-            help="Download from autoland",
-        )
-
-        build_group = self.parser.add_argument_group("Build Arguments")
-        build_group.add_argument(
-            "-d",
-            "--debug",
-            action="store_true",
-            help="Get debug builds w/ symbols (default=optimized).",
-        )
-        build_group.add_argument(
-            "-a",
-            "--asan",
-            action="store_true",
-            help="Download AddressSanitizer builds.",
-        )
-        build_group.add_argument(
-            "-t", "--tsan", action="store_true", help="Download ThreadSanitizer builds."
-        )
-        build_group.add_argument(
-            "--fuzzing", action="store_true", help="Download --enable-fuzzing builds."
-        )
-        build_group.add_argument(
-            "--fuzzilli",
-            action="store_true",
-            help="Download --enable-js-fuzzilli builds.",
-        )
-        build_group.add_argument(
-            "--coverage", action="store_true", help="Download --coverage builds."
-        )
-        build_group.add_argument(
-            "--valgrind", action="store_true", help="Download Valgrind builds."
-        )
-        build_group.add_argument(
-            "--no-opt", action="store_true", help="Download non-optimized builds."
-        )
-
-        self.parser.add_argument(
-            "--gtest",
-            action="store_true",
-            help=SUPPRESS,
-        )
-
-        misc_group = self.parser.add_argument_group("Misc. Arguments")
-        misc_group.add_argument("-n", "--name", help="Specify a name (default=auto)")
-        misc_group.add_argument(
-            "-o",
-            "--out",
-            default=os.getcwd(),
-            help="Specify output directory (default=.)",
-        )
-        misc_group.add_argument(
-            "--dry-run",
-            action="store_true",
-            help="Search for build and output metadata only, don't download anything.",
-        )
-
-        near_group = self.parser.add_argument_group(
-            "Near Arguments",
-            "If the specified build isn't found, iterate over "
-            "builds in the specified direction",
-        )
-        near_args = near_group.add_mutually_exclusive_group()
-        near_args.add_argument(
-            "--nearest-newer",
-            action="store_const",
-            const=BuildSearchOrder.ASC,
-            dest="nearest",
-            help="Search from specified build in ascending order",
-        )
-        near_args.add_argument(
-            "--nearest-older",
-            action="store_const",
-            const=BuildSearchOrder.DESC,
-            dest="nearest",
-            help="Search from the specified build in descending order",
-        )
-
-    @staticmethod
-    def is_build_ns(build_id: str) -> bool:
-        """Check if supplied build_id is a namespace
-
-        Arguments:
-            build_id: Build identifier to check
-        """
-        return (
-            re.match(
-                r"(\d{4}-\d{2}-\d{2}|[0-9A-Fa-f]{12}|[0-9A-Fa-f]{40}|latest)$", build_id
-            )
-            is None
-        )
-
-    def sanity_check(self, args: Namespace) -> None:
-        """Perform parser checks
-
-        Arguments:
-            args: Parsed arguments
-        """
-        # multiple-inheritance support
-        if hasattr(super(), "sanity_check"):
-            # pylint: disable=no-member
-            super().sanity_check(args)  # type: ignore
-
-        if self.is_build_ns(args.build):
-            # this is a custom build
-            # ensure conflicting options are not set
-            if args.branch is not None:
-                self.parser.error(
-                    "Cannot specify --build namespace and branch argument: "
-                    f"{args.branch}"
-                )
-            if args.debug:
-                self.parser.error("Cannot specify --build namespace and --debug")
-            if args.asan:
-                self.parser.error("Cannot specify --build namespace and --asan")
-            if args.tsan:
-                self.parser.error("Cannot specify --build namespace and --tsan")
-            if args.fuzzing:
-                self.parser.error("Cannot specify --build namespace and --fuzzing")
-            if args.coverage:
-                self.parser.error("Cannot specify --build namespace and --coverage")
-            if args.valgrind:
-                self.parser.error("Cannot specify --build namespace and --valgrind")
-            if args.no_opt:
-                self.parser.error("Cannot specify --build namespace and --no-opt")
-
-        if args.gtest:
-            LOG.warning("--gtest is deprecated, add 'gtest' to --target instead.")
-            args.target.append("gtest")
-
-        if "firefox" in args.target and args.fuzzilli:
-            self.parser.error("Cannot specify --target firefox and --fuzzilli")
-
-    def parse_args(self, argv: Optional[Sequence[str]] = None) -> Namespace:
-        """Parse and validate args
-
-        Arguments:
-            argv: a list of arguments
-        """
-        args = self.parser.parse_args(argv)
-        self.sanity_check(args)
-        return args
-
-
-class BuildSearchOrder(Enum):
-    """Search direction when searching for "nearest" builds"""
-
-    ASC = 1
-    DESC = 2
 
 
 class Fetcher:
